+++ src/Traits/WithData.php
@@ -214,4 +214,17 @@ trait WithData
 
         return $currentTableAlias.'_'.$relationPart;
     }
+
+    public function renderingWithData($view, $data)
+    {
+        $view = $view->with(
+            [
+                'filterGenericData' => $this->getFilterGenericData(),
+                'columns' => $this->getColumns(),
+                'rows' => $this->getRows(),
+                'customView' => $this->customView(),
+            ]
+        );
+    }
+
 }